add endpoint arg to client, fix in startup function
